Update DBAccess and JanusEventsFactory
